Make main.rb to contain only the user choice
